Configuring Advanced Network Security with OPNsense Sensei
This project accomplished the configuration of Opnsense Sensei to specifically monitor and control the Network, allowing selective access websites like Google while restricting access to certain websites. Additionally, it involved adjusting the security posture of the Web Filter and modifying the dashboard for enhanced monitoring, including the tracking of blocked hosts and activities.